requirements.txt
Django==5.0.6
Pillow==10.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-imagekit==5.0.0
django-filter==24.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-leaflet==0.30.1
requests==2.31.0
crispy-tailwind==1.0.3
django-tinymce==4.0.0
